Move request and response to log functions into separate file
